Refactor playback and repository code to remove null checks on items, improve logging, and enhance UI components
- EnhancedPlaybackManager: Simplified item ID retrieval by removing null check.
- JellyfinMediaRepository: Removed null checks for items in multiple methods to streamline data handling.
- JellyfinRepository: Updated item retrieval methods to eliminate null checks and improve logging for debugging.
- JellyfinSearchRepository: Simplified item retrieval by removing null checks.
- JellyfinUserRepository: Cleaned up user policy checks for better readability.
- LibraryHealthIndicator: Updated health status handling to include UNKNOWN state.
- OptimizedImageLoader: Fixed potential null pointer by ensuring type is not nullable.
- TvAudioVisualizer: Adjusted waveform drawing logic for better performance.
- themes.xml: Added Material 3 aligned theme for MediaRouter.

+++ app/src/main/java/com/rpeters/jellyfin/ui/player/tv/TvAudioVisualizer.kt
@@ -225,9 +225,9 @@ private fun DrawScope.drawWaveform(
                 )
 
             if (i == 0) {
-                moveTo(x, y.toFloat())
+                moveTo(x, y)
             } else {
-                lineTo(x, y.toFloat())
+                lineTo(x, y)
             }
         }
     }
